requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
beautifulsoup4==4.12.3
billiard==4.2.0
bs4==0.0.2
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
django-celery-beat==2.6.0
django-cors-headers==4.3.1
django-cron==0.6.0
django-crontab==0.7.1
django-extensions==3.2.3
django-sslserver==0.22
django-timezone-field==6.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
filelock==3.13.1
fsspec==2024.2.0
google_search_results==2.4.2
gunicorn==22.0.0
huggingface-hub==0.22.2
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.3
kombu==5.3.7
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
mysqlclient==2.2.4
networkx==3.2.1
numpy==1.26.3
openpyxl==3.1.2
packaging==24.0
pandas==2.2.2
pillow==10.2.0
prompt_toolkit==3.0.47
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.1.2
python-crontab==3.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
regex==2024.4.28
requests==2.31.0
requests-file==2.1.0
safetensors==0.4.3
sendotp==0.0.4
six==1.16.0
soupsieve==2.5
sqlparse==0.5.0
sympy==1.12
tbb==2021.11.0
tldextract==5.1.2
tokenizers==0.19.1
torch==2.3.0+cpu
torchaudio==2.3.0+cpu
torchvision==0.18.0+cpu
tqdm==4.66.2
transformers==4.40.1
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.7.0
xlrd==2.0.1
